Hide Firestore codec class and move each class into separate files (#452)
