+++ packages/cloud_firestore/lib/cloud_firestore.dart
@@ -20,8 +20,11 @@ part 'src/document_change.dart';
 part 'src/document_snapshot.dart';
 part 'src/document_reference.dart';
 part 'src/firestore.dart';
+part 'src/geo_point.dart';
 part 'src/query.dart';
 part 'src/query_snapshot.dart';
 part 'src/set_options.dart';
+part 'src/firestore_message_codec.dart';
 part 'src/snapshot_metadata.dart';
+part 'src/transaction.dart';
 part 'src/write_batch.dart';
