+++ packages/cloud_firestore/lib/src/firestore.dart
@@ -106,108 +106,3 @@ class Firestore {
     return result?.cast<String, dynamic>() ?? <String, dynamic>{};
   }
 }
-
-typedef Future<dynamic> TransactionHandler(Transaction transaction);
-
-class Transaction {
-  int _transactionId;
-
-  Transaction(this._transactionId);
-
-  Future<DocumentSnapshot> get(DocumentReference documentReference) async {
-    final dynamic result = await Firestore.channel
-        .invokeMethod('Transaction#get', <String, dynamic>{
-      'transactionId': _transactionId,
-      'path': documentReference.path,
-    });
-    if (result != null) {
-      return new DocumentSnapshot._(documentReference.path,
-          result['data'].cast<String, dynamic>(), Firestore.instance);
-    } else {
-      return null;
-    }
-  }
-
-  Future<void> delete(DocumentReference documentReference) async {
-    return Firestore.channel
-        .invokeMethod('Transaction#delete', <String, dynamic>{
-      'transactionId': _transactionId,
-      'path': documentReference.path,
-    });
-  }
-
-  Future<void> update(
-      DocumentReference documentReference, Map<String, dynamic> data) async {
-    return Firestore.channel
-        .invokeMethod('Transaction#update', <String, dynamic>{
-      'transactionId': _transactionId,
-      'path': documentReference.path,
-      'data': data,
-    });
-  }
-
-  Future<void> set(
-      DocumentReference documentReference, Map<String, dynamic> data) async {
-    return Firestore.channel.invokeMethod('Transaction#set', <String, dynamic>{
-      'transactionId': _transactionId,
-      'path': documentReference.path,
-      'data': data,
-    });
-  }
-}
-
-class FirestoreMessageCodec extends StandardMessageCodec {
-  const FirestoreMessageCodec();
-
-  static const int _kDateTime = 128;
-  static const int _kGeoPoint = 129;
-  static const int _kDocumentReference = 130;
-
-  @override
-  void writeValue(WriteBuffer buffer, dynamic value) {
-    if (value is DateTime) {
-      buffer.putUint8(_kDateTime);
-      buffer.putInt64(value.millisecondsSinceEpoch);
-    } else if (value is GeoPoint) {
-      buffer.putUint8(_kGeoPoint);
-      buffer.putFloat64(value.latitude);
-      buffer.putFloat64(value.longitude);
-    } else if (value is DocumentReference) {
-      buffer.putUint8(_kDocumentReference);
-      final List<int> bytes = utf8.encoder.convert(value.path);
-      writeSize(buffer, bytes.length);
-      buffer.putUint8List(bytes);
-    } else {
-      super.writeValue(buffer, value);
-    }
-  }
-
-  @override
-  dynamic readValueOfType(int type, ReadBuffer buffer) {
-    switch (type) {
-      case _kDateTime:
-        return new DateTime.fromMillisecondsSinceEpoch(buffer.getInt64());
-      case _kGeoPoint:
-        return new GeoPoint(buffer.getFloat64(), buffer.getFloat64());
-      case _kDocumentReference:
-        final int length = readSize(buffer);
-        final String path = utf8.decoder.convert(buffer.getUint8List(length));
-        return Firestore.instance.document(path);
-      default:
-        return super.readValueOfType(type, buffer);
-    }
-  }
-}
-
-class GeoPoint {
-  final double latitude;
-  final double longitude;
-  const GeoPoint(this.latitude, this.longitude);
-
-  @override
-  bool operator ==(dynamic o) =>
-      o is GeoPoint && o.latitude == latitude && o.longitude == longitude;
-
-  @override
-  int get hashCode => hashValues(latitude, longitude);
-}
